requirements.txt
asgiref==3.2.7
cycler==0.10.0
Django==3.0.6
imutils==0.5.3
kiwisolver==1.2.0
matplotlib==3.2.1
numpy==1.18.4
opencv-contrib-python==4.2.0.34
opencv-python==4.2.0.34
pandas==1.0.3
Pillow==7.1.2
pyparsing==2.4.7
python-dateutil==2.8.1
pytz==2020.1
six==1.14.0
sqlparse==0.3.1
django-cleanup==4.0.0
django-multiselectfield==0.1.12
SQLAlchemy==1.3.17
django-mail-templated==2.6.5
